validation for `// @tstyche template`
