--- tests/__fixtures__/validation-template/__typetests__/missing-export.tst.ts
@@ -1,14 +1,0 @@
-// @tstyche template
-
-let testText = `
-import { expect, test } from "tstyche";
-`;
-
-for (const target of ["string", "number"]) {
-  testText += `test("is ${target} a string?", () => {
-  expect<string>().type.toBe<${target}>();
-});
-`;
-}
-
-export { testText };
