--- tests/__fixtures__/validation-template/__typetests__/test-text-syntax-errors.tst.ts
@@ -1,22 +1,0 @@
-// @tstyche template
-
-const testText = `import { expect, test } from "tstyche";
-
-declare function one(a: string): void;
-
-test("is syntax error?", () => {
-  one(());
-});
-
-test("is syntax error?", () => {
-  one(
-});
-
-test("is skipped?", () => {
-  expect(one("abc")).type.toBe<void>();
-});
-
-test("is broken?"
-`;
-
-export default testText;
