--- tests/__fixtures__/validation-template/__typetests__/test-text-type-errors.tst.ts
@@ -1,14 +1,0 @@
-// @tstyche template
-
-let testText = `
-import { expect, test } from "tstyche";
-`;
-
-for (const target of ["string", "number"]) {
-  testText += `test("is ${target} a string?", () => {
-  expect<string>().toBe<${target}>();
-});
-`;
-}
-
-export default testText;
