--- tests/__fixtures__/validation-template/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "include": ["**/*"]
-}
