+++ tests/validation-template.test.js
@@ -5,9 +5,9 @@ import { normalizeOutput } from "./__utilities__/output.js";
 import { spawnTyche } from "./__utilities__/tstyche.js";
 
 const testFileName = getTestFileName(import.meta.url);
-const fixtureUrl = getFixtureFileUrl(testFileName);
+const fixtureUrl = getFixtureFileUrl(testFileName, { generated: true });
 
-await test("template test file", async (t) => {
+await test("'// @tstyche template' directive", async (t) => {
   // TODO remove this check after dropping support for Node.js 20
   if (process.versions.node.startsWith("20")) {
     t.skip();
@@ -15,8 +15,68 @@ await test("template test file", async (t) => {
     return;
   }
 
+  t.afterEach(async () => {
+    await clearFixture(fixtureUrl);
+  });
+
+  await t.test("does not take an argument", async () => {
+    const testFileText = `// @tstyche template nope
+
+let testText = \`import { expect, test } from "tstyche";
+\`;
+
+for (const target of ["string", "number"]) {
+  testText += \`test("is \${target} a string?", () => {
+  expect<string>().type.toBe<\${target}>();
+});
+\`;
+}
+
+export default testText;
+`;
+
+    await writeFixture(fixtureUrl, {
+      ["__typetests__/sample.tst.ts"]: testFileText,
+    });
+
+    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, [], {
+      env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
+    });
+
+    await assert.matchSnapshot(stderr, {
+      fileName: `${testFileName}-does-not-take-argument-stderr`,
+      testFileUrl: import.meta.url,
+    });
+
+    await assert.matchSnapshot(normalizeOutput(stdout), {
+      fileName: `${testFileName}-does-not-take-argument-stdout`,
+      testFileUrl: import.meta.url,
+    });
+
+    assert.equal(exitCode, 1);
+  });
+
   await t.test("must export a string", async () => {
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, ["must-export-a-string"], {
+    const testFileText = `// @tstyche template
+
+let testText = \`import { expect, test } from "tstyche";
+\`;
+
+for (const target of ["string", "number"]) {
+  testText += \`test("is \${target} a string?", () => {
+  expect<string>().type.toBe<\${target}>();
+});
+\`;
+}
+
+export default () => testText;
+`;
+
+    await writeFixture(fixtureUrl, {
+      ["__typetests__/sample.tst.ts"]: testFileText,
+    });
+
+    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, [], {
       env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
     });
 
@@ -34,7 +94,26 @@ await test("template test file", async (t) => {
   });
 
   await t.test("handles missing export", async () => {
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, ["missing-export"], {
+    const testFileText = `// @tstyche template
+
+let testText = \`import { expect, test } from "tstyche";
+\`;
+
+for (const target of ["string", "number"]) {
+  testText += \`test("is \${target} a string?", () => {
+  expect<string>().type.toBe<\${target}>();
+});
+\`;
+}
+
+export { testText };
+`;
+
+    await writeFixture(fixtureUrl, {
+      ["__typetests__/sample.tst.ts"]: testFileText,
+    });
+
+    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, [], {
       env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
     });
 
@@ -52,7 +131,7 @@ await test("template test file", async (t) => {
   });
 
   await t.test("handles test file syntax errors", async () => {
-    const templateText = `// @tstyche template
+    const testFileText = `// @tstyche template
 
 function getTestText() {
   return \`import { expect, test } from "tstyche";
@@ -63,23 +142,12 @@ expect<string>().type.toBe<string>();
 export default getTestText(;
 `;
 
-    const fixtureUrl = getFixtureFileUrl(testFileName, { generated: true });
-
-    t.after(async () => {
-      await clearFixture(fixtureUrl);
-    });
-
     await writeFixture(fixtureUrl, {
-      ["__typetests__/template.tst.ts"]: templateText,
+      ["__typetests__/sample.tst.ts"]: testFileText,
     });
 
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, ["template"], {
+    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, [], {
       env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
-    });
-
-    await assert.matchSnapshot(normalizeOutput(stdout), {
-      fileName: `${testFileName}-test-file-syntax-errors-stdout`,
-      testFileUrl: import.meta.url,
     });
 
     await assert.matchSnapshot(stderr, {
@@ -87,17 +155,38 @@ export default getTestText(;
       testFileUrl: import.meta.url,
     });
 
+    await assert.matchSnapshot(normalizeOutput(stdout), {
+      fileName: `${testFileName}-test-file-syntax-errors-stdout`,
+      testFileUrl: import.meta.url,
+    });
+
     assert.equal(exitCode, 1);
   });
 
   await t.test("handles test file type errors", async () => {
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, ["test-file-type"], {
-      env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
+    const testFileText = `// @tstyche template
+
+let testText = \`import { expect, test } from "tstyche";
+\`;
+
+export const a: number = "nine";
+
+for (const target of ["string", "number"]) {
+  testText += \`test("is \${target} a string?", () => {
+  expect<string>().type.toBe<\${target}>();
+});
+\`;
+}
+
+export default testText;
+`;
+
+    await writeFixture(fixtureUrl, {
+      ["__typetests__/sample.tst.ts"]: testFileText,
     });
 
-    await assert.matchSnapshot(normalizeOutput(stdout), {
-      fileName: `${testFileName}-test-file-type-errors-stdout`,
-      testFileUrl: import.meta.url,
+    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, [], {
+      env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
     });
 
     await assert.matchSnapshot(stderr, {
@@ -105,17 +194,45 @@ export default getTestText(;
       testFileUrl: import.meta.url,
     });
 
+    await assert.matchSnapshot(normalizeOutput(stdout), {
+      fileName: `${testFileName}-test-file-type-errors-stdout`,
+      testFileUrl: import.meta.url,
+    });
+
     assert.equal(exitCode, 1);
   });
 
   await t.test("handles test text syntax errors", async () => {
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, ["test-text-syntax"], {
-      env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
+    const testFileText = `// @tstyche template
+
+const testText = \`import { expect, test } from "tstyche";
+
+declare function one(a: string): void;
+
+test("is syntax error?", () => {
+  one(());
+});
+
+test("is syntax error?", () => {
+  one(
+});
+
+test("is skipped?", () => {
+  expect(one("abc")).type.toBe<void>();
+});
+
+test("is broken?"
+\`;
+
+export default testText;
+`;
+
+    await writeFixture(fixtureUrl, {
+      ["__typetests__/sample.tst.ts"]: testFileText,
     });
 
-    await assert.matchSnapshot(normalizeOutput(stdout), {
-      fileName: `${testFileName}-test-text-syntax-errors-stdout`,
-      testFileUrl: import.meta.url,
+    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, [], {
+      env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
     });
 
     await assert.matchSnapshot(stderr, {
@@ -123,21 +240,45 @@ export default getTestText(;
       testFileUrl: import.meta.url,
     });
 
+    await assert.matchSnapshot(normalizeOutput(stdout), {
+      fileName: `${testFileName}-test-text-syntax-errors-stdout`,
+      testFileUrl: import.meta.url,
+    });
+
     assert.equal(exitCode, 1);
   });
 
   await t.test("handles test text type errors", async () => {
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, ["test-text-type"], {
-      env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
+    const testFileText = `// @tstyche template
+
+let testText = \`import { expect, test } from "tstyche";
+\`;
+
+for (const target of ["string", "number"]) {
+  testText += \`test("is \${target} a string?", () => {
+  expect<string>().toBe<\${target}>();
+});
+\`;
+}
+
+export default testText;
+`;
+
+    await writeFixture(fixtureUrl, {
+      ["__typetests__/sample.tst.ts"]: testFileText,
     });
 
-    await assert.matchSnapshot(normalizeOutput(stdout), {
-      fileName: `${testFileName}-test-text-type-errors-stdout`,
-      testFileUrl: import.meta.url,
+    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl, [], {
+      env: { ["NODE_OPTIONS"]: "--experimental-strip-types --no-warnings" },
     });
 
     await assert.matchSnapshot(stderr, {
       fileName: `${testFileName}-test-text-type-errors-stderr`,
+      testFileUrl: import.meta.url,
+    });
+
+    await assert.matchSnapshot(normalizeOutput(stdout), {
+      fileName: `${testFileName}-test-text-type-errors-stdout`,
       testFileUrl: import.meta.url,
     });
